fix: extract maybeGunzip into GzipUtils to avoid Jenkins class init in unit tests
BitBucketPPRHookReceiver has a static field initializer that calls
BitBucketPPRPluginConfig.getInstance(), which requires a running
Jenkins instance. Calling any static method on the class triggers
class loading and throws ExceptionInInitializerError in plain unit
tests.

Extract maybeGunzip() into a pure-Java GzipUtils helper class with
no Jenkins dependencies. Tests call GzipUtils.maybeGunzip() directly
and compile/run without a Jenkins harness.
